refactor: convert common module to kotlin
