+++ bro-common/build.gradle.kts
@@ -1,11 +1,13 @@
 import me.xx2bab.bro.build.BuildConfig
 
 plugins {
-    id("java-library")
+    id("kotlin")
 }
 
 dependencies {
     implementation(fileTree(mapOf("dir" to "libs", "include" to arrayOf("*.jar"))))
+    implementation(kotlin("stdlib-jdk8"))
+    implementation(kotlin("reflect"))
     compileOnly(BuildConfig.Deps.androidRuntime)
     compileOnly(BuildConfig.Deps.orgJson)
     api(BuildConfig.Deps.supportAnno)
