--- bro-common/src/main/java/me/xx2bab/bro/common/Constants.java
@@ -1,22 +1,0 @@
-package me.xx2bab.bro.common;
-
-
-public class Constants {
-
-    public static final String ANNO_PROC_ARG_MODULE_NAME = "bro.moduleName";
-    public static final String ANNO_PROC_ARG_MODULE_BUILD_TYPE = "bro.moduleBuildType";
-    public static final String ANNO_PROC_ARG_MODULE_BUILD_DIR = "bro.moduleBroBuildDir";
-    public static final String ANNO_PROC_ARG_MODULE_PROCESSOR_CLASSES = "bro.moduleProcessorClasses";
-    public static final String ANNO_PROC_ARG_MODULE_GENERATOR_CLASSPATHS = "bro.moduleProcessorClassPaths";
-    public static final String ANNO_PROC_ARG_APP_PACKAGE_NAME = "bro.appPackageName";
-    public static final String ANNO_PROC_ARG_APP_META_DATA_INPUT_PATH = "bro.appMetaDataInputPath";
-    public static final String ANNO_PROC_ARG_APP_APT_PATH = "bro.appAptPath";
-    public static final String ANNO_PROC_ARG_LIB_META_DATA_OUTPUT_PATH = "bro.libMetaDataOutputPath";
-
-    public static final String MODULE_META_INFO_FILE_SUFFIX = ".bro";
-    public static final String MODULE_COMPILER_OPTIONS_FILE = "annotation_processor_options.json";
-
-    public static final String GEN_PACKAGE_NAME = "me.xx2bab.bro.core";
-    public static final String GEN_CLASS_SUFFIX = "Impl";
-
-}
