--- bro-common/src/main/java/me/xx2bab/bro/common/IBroApi.java
@@ -1,7 +1,0 @@
-package me.xx2bab.bro.common;
-
-public interface IBroApi {
-
-    void onCreate();
-
-}
