--- bro-common/src/main/java/me/xx2bab/bro/common/IBroModule.java
@@ -1,13 +1,0 @@
-package me.xx2bab.bro.common;
-
-import android.content.Context;
-
-import java.util.Set;
-
-public interface IBroModule {
-
-    Set<Class<? extends IBroApi>> getLaunchDependencies();
-
-    void onCreate(Context context);
-
-}
