--- bro-common/src/main/java/me/xx2bab/bro/common/ModuleType.java
@@ -1,8 +1,0 @@
-package me.xx2bab.bro.common;
-
-public enum ModuleType {
-
-    APPLICATION,
-    LIBRARY
-    
-}
