--- bro-common/src/main/java/me/xx2bab/bro/common/gen/GenOutputs.java
@@ -1,30 +1,0 @@
-package me.xx2bab.bro.common.gen;
-
-import java.io.File;
-import java.util.Arrays;
-
-import me.xx2bab.bro.common.Constants;
-
-public class GenOutputs {
-
-    // The package name of the app.
-    public String appPackageName;
-
-    // The directory for placing new java class generated by the generator.
-    public File appAptGenDirectory;
-
-    // The directory for placing other building artifacts like docs, etc.
-    public File broBuildDirectory;
-
-    public String generateClassNameForImplementation(Class clazz) {
-        String[] className = clazz.getCanonicalName().split("\\.");
-        StringBuilder builder = new StringBuilder();
-        builder.append(className[0]);
-        for (int i = 1; i < className.length; i++) {
-            builder.append(className[i].substring(0, 1).toUpperCase())
-                    .append(className[i].substring(1));
-        }
-        return builder.toString() + Constants.GEN_CLASS_SUFFIX;
-    }
-
-}
