--- bro-common/src/main/java/me/xx2bab/bro/common/gen/anno/AnnotatedElement.java
@@ -1,16 +1,0 @@
-package me.xx2bab.bro.common.gen.anno;
-
-import java.lang.annotation.ElementType;
-import java.util.Set;
-
-/**
- * Created on 2019-07-17
- */
-public class AnnotatedElement {
-
-    public String name;
-    public ElementType type;
-    public String clazz;
-    public Set<Annotation> annotations;
-
-}
