--- bro-common/src/main/java/me/xx2bab/bro/common/gen/anno/Annotation.java
@@ -1,10 +1,0 @@
-package me.xx2bab.bro.common.gen.anno;
-
-import java.util.TreeMap;
-
-public class Annotation {
-
-    public String name;
-    public TreeMap<String, String> values;
-
-}
