--- bro-common/src/main/java/me/xx2bab/bro/common/gen/anno/IBroApiInterfaceAndAliasMap.java
@@ -1,7 +1,0 @@
-package me.xx2bab.bro.common.gen.anno;
-
-public interface IBroApiInterfaceAndAliasMap {
-
-    String getAliasByInterface(String interfaze);
-
-}
