+++ bro-common/src/main/kotlin/me/xx2bab/bro/common/Constants.kt
@@ -1,0 +1,17 @@
+package me.xx2bab.bro.common
+
+object Constants {
+    const val ANNO_PROC_ARG_MODULE_NAME = "bro.moduleName"
+    const val ANNO_PROC_ARG_MODULE_BUILD_TYPE = "bro.moduleBuildType"
+    const val ANNO_PROC_ARG_MODULE_BUILD_DIR = "bro.moduleBroBuildDir"
+    const val ANNO_PROC_ARG_MODULE_PROCESSOR_CLASSES = "bro.moduleProcessorClasses"
+    const val ANNO_PROC_ARG_MODULE_GENERATOR_CLASSPATHS = "bro.moduleProcessorClassPaths"
+    const val ANNO_PROC_ARG_APP_PACKAGE_NAME = "bro.appPackageName"
+    const val ANNO_PROC_ARG_APP_META_DATA_INPUT_PATH = "bro.appMetaDataInputPath"
+    const val ANNO_PROC_ARG_APP_APT_PATH = "bro.appAptPath"
+    const val ANNO_PROC_ARG_LIB_META_DATA_OUTPUT_PATH = "bro.libMetaDataOutputPath"
+    const val MODULE_META_INFO_FILE_SUFFIX = ".bro"
+    const val MODULE_COMPILER_OPTIONS_FILE = "annotation_processor_options.json"
+    const val GEN_PACKAGE_NAME = "me.xx2bab.bro.core"
+    const val GEN_CLASS_SUFFIX = "Impl"
+}
