+++ bro-common/src/main/kotlin/me/xx2bab/bro/common/IBroApi.kt
@@ -1,0 +1,5 @@
+package me.xx2bab.bro.common
+
+interface IBroApi {
+    fun onCreate()
+}
