+++ bro-common/src/main/kotlin/me/xx2bab/bro/common/IBroModule.kt
@@ -1,0 +1,11 @@
+package me.xx2bab.bro.common
+
+import android.content.Context
+
+interface IBroModule {
+
+    fun getLaunchDependencies(): Set<Class<out IBroApi>>?
+
+    fun onCreate(context: Context?)
+
+}
