+++ bro-common/src/main/kotlin/me/xx2bab/bro/common/ModuleType.kt
@@ -1,0 +1,5 @@
+package me.xx2bab.bro.common
+
+enum class ModuleType {
+    APPLICATION, LIBRARY
+}
