+++ bro-common/src/main/kotlin/me/xx2bab/bro/common/gen/anno/Annotation.kt
@@ -1,0 +1,5 @@
+package me.xx2bab.bro.common.gen.anno
+
+import java.util.*
+
+data class Annotation(val name: String, val values: TreeMap<String, String>)
