+++ bro-common/src/main/kotlin/me/xx2bab/bro/common/gen/anno/IBroAliasRoutingTable.kt
@@ -1,0 +1,9 @@
+package me.xx2bab.bro.common.gen.anno
+
+import me.xx2bab.bro.common.BroProperties
+
+interface IBroAliasRoutingTable {
+    fun getRoutingMapByAnnotation(
+            annotation: Class<out kotlin.Annotation>
+    ): Map<String?, BroProperties?>?
+}
