+++ bro-common/src/main/kotlin/me/xx2bab/bro/common/gen/anno/IBroApiInterfaceAndAliasMap.kt
@@ -1,0 +1,5 @@
+package me.xx2bab.bro.common.gen.anno
+
+interface IBroApiInterfaceAndAliasMap {
+    fun getAliasByInterface(interfaze: String?): String?
+}
