+++ sample/build.gradle
@@ -6,7 +6,7 @@ buildscript {
     }
 
     ext {
-        broLatestReleaseVersion = "1.3.3"
+        broLatestReleaseVersion = "1.3.4.3-SNAPSHOT"
         broSourceCompatibilityVersion = JavaVersion.VERSION_1_7
         broTargetCompatibilityVersion = JavaVersion.VERSION_1_7
     }
